requirements2.txt
aiofiles==24.1.0
aiohappyeyeballs==2.6.1
aiohttp==3.11.14
aiosignal==1.3.2
aiosqlite==0.21.0
altair==5.5.0
annotated-types==0.7.0
anyio==4.9.0
attrs==25.3.0
azure-ai-textanalytics==5.3.0
azure-common==1.1.28
azure-core==1.33.0
azure-identity==1.21.0
beautifulsoup4==4.13.3
blinker==1.9.0
cachetools==5.5.2
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
cryptography==44.0.2
cssselect==1.3.0
dill==0.4.0
distro==1.9.0
docker==7.1.0
docling==2.28.4
docling-core==2.25.0
docling-ibm-models==3.4.1
docling-parse==4.0.0
duckduckgo_search==8.0.1
easyocr==1.7.2
et_xmlfile==2.0.0
fake-http-header==0.3.5
fake-useragent==2.2.0
faust-cchardet==2.1.19
filelock==3.18.0
filetype==1.2.0
frozenlist==1.6.0
fsspec==2025.3.2
gitdb==4.0.12
GitPython==3.1.44
greenlet==3.2.0
h11==0.14.0
httpcore==1.0.8
httpx==0.28.1
huggingface-hub==0.30.2
humanize==4.12.2
idna==3.10
imageio==2.37.0
importlib_metadata==8.6.1
inscriptis==2.6.0
isodate==0.7.2
Jinja2==3.1.6
jiter==0.9.0
joblib==1.4.2
jsonlines==3.1.0
jsonref==1.1.0
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
latex2mathml==3.77.0
lazy_loader==0.4
litellm==1.66.3
llm-reader @ git+https://github.com/m92vyas/llm-reader.git@182987e3afd54956a54719cae7b21e9a719ca5fe
lxml==5.3.2
markdown-it-py==3.0.0
markdownify==1.1.0
marko==2.1.3
MarkupSafe==3.0.2
mdurl==0.1.2
minify_html==0.16.4
mpire==2.10.2
mpmath==1.3.0
msal==1.32.0
msal-extensions==1.3.1
multidict==6.4.3
multiprocess==0.70.17
narwhals==1.35.0
networkx==3.4.2
ninja==1.11.1.4
nltk==3.9.1
numpy==2.2.4
nvidia-cublas-cu12==12.4.5.8
nvidia-cuda-cupti-cu12==12.4.127
nvidia-cuda-nvrtc-cu12==12.4.127
nvidia-cuda-runtime-cu12==12.4.127
nvidia-cudnn-cu12==9.1.0.70
nvidia-cufft-cu12==11.2.1.3
nvidia-curand-cu12==10.3.5.147
nvidia-cusolver-cu12==11.6.1.9
nvidia-cusparse-cu12==12.3.1.170
nvidia-cusparselt-cu12==0.6.2
nvidia-nccl-cu12==2.21.5
nvidia-nvjitlink-cu12==12.4.127
nvidia-nvtx-cu12==12.4.127
openai==1.68.2
opencv-python-headless==4.11.0.86
openpyxl==3.1.5
outcome==1.3.0.post0
packaging==24.2
pandas==2.2.3
pgvector==0.4.0
pillow==11.2.1
playwright==1.51.0
pluggy==1.5.0
primp==0.15.0
propcache==0.3.1
protobuf==5.29.4
psutil==7.0.0
psycopg==3.2.6
psycopg-binary==3.2.6
pyarrow==19.0.1
pyclipper==1.3.0.post6
pycparser==2.22
pydantic==2.11.3
pydantic-settings==2.9.1
pydantic_core==2.33.1
pydeck==0.9.1
pyee==12.1.1
Pygments==2.19.1
PyJWT==2.10.1
pylatexenc==2.10
pyOpenSSL==25.0.0
pypdf==3.1.0
pypdfium2==4.30.1
pyperclip==1.9.0
PySocks==1.7.1
python-bidi==0.6.6
python-dateutil==2.9.0.post0
python-docx==1.1.2
python-dotenv==1.1.0
python-pptx==1.0.2
pytz==2025.2
PyYAML==6.0.2
rank-bm25==0.2.2
referencing==0.36.2
regex==2024.11.6
requests==2.32.3
rich==14.0.0
rpds-py==0.24.0
rtree==1.4.0
safetensors==0.5.3
scikit-image==0.25.2
scikit-learn==1.6.1
scipy==1.15.2
selenium==4.30.0
semchunk==2.2.2
sentence-transformers==4.0.1
setuptools==78.1.0
shapely==2.1.0
shellingham==1.5.4
six==1.17.0
smmap==5.0.2
smolagents==1.14.0
sniffio==1.3.1
snowballstemmer==2.2.0
sortedcontainers==2.4.0
soupsieve==2.6
SQLAlchemy==2.0.40
sqlglot==26.16.4
streamlit==1.44.1
sympy==1.13.1
tabulate==0.9.0
tenacity==9.1.2
tf-playwright-stealth==1.1.2
threadpoolctl==3.6.0
tifffile==2025.3.30
tiktoken==0.9.0
tokenizers==0.21.1
toml==0.10.2
torch==2.6.0
torchvision==0.21.0
tornado==6.4.2
tqdm==4.67.1
transformers==4.50.3
trio==0.29.0
trio-websocket==0.12.2
triton==3.2.0
typer==0.12.5
typing-inspection==0.4.0
typing_extensions==4.13.2
tzdata==2025.2
urllib3==2.4.0
watchdog==6.0.0
websocket-client==1.8.0
wsproto==1.2.0
XlsxWriter==3.2.3
xxhash==3.5.0
yarl==1.20.0
zipp==3.21.0
